Remove unnecessary entities; reorganize remaining.
Resolves #113
