+++ client.src/core/entities/user.js
