+++ client.src/modules/profile/entities/github-user.js
@@ -4,7 +4,7 @@
 //
 
 import * as bb from 'backbone';
-import githubApiHelpers from '../../helpers/github-api-helpers';
+import githubApiHelpers from '../../../helpers/github-api-helpers';
 
 export default bb.Model.extend({
   urlRoot: function() {
