--- client.src/shared/entities/gistbook.js
@@ -1,13 +1,0 @@
-//
-// gistbook
-// A Model for a Gistbook
-//
-
-import * as bb from 'backbone';
-import githubApiHelpers from '../../helpers/github-api-helpers';
-
-export default bb.Model.extend({
-  urlRoot: function() {
-    return githubApiHelpers.url + '/gists';
-  }
-});
